requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
crewai==0.51.1
crewai-tools==0.1.0
pydantic==2.9.2
httpx==0.27.0
openai==1.51.0
python-dotenv==1.0.1
